requirements.txt
pillow==10.0.0
requests==2.31.0